Misc persistence fixes.
- Tests
- Forward Started message to actor
- Use ulong instead of int for event index

+++ src/Proto.Persistence/IProviderState.cs
@@ -12,11 +12,11 @@ namespace Proto.Persistence
 {
     public interface IProviderState
     {
-        Task GetEventsAsync(string actorName, int eventIndexStart, Action<object> callback);
-        Task<Tuple<object, int>> GetSnapshotAsync(string actorName);
+        Task GetEventsAsync(string actorName, ulong eventIndexStart, Action<object> callback);
+        Task<Tuple<object, ulong>> GetSnapshotAsync(string actorName);
         int GetSnapshotInterval();
-        Task PersistEventAsync(string actorName, int eventIndex, IMessage @event);
-        Task PersistSnapshotAsync(string actorName, int eventIndex, IMessage snapshot);
+        Task PersistEventAsync(string actorName, ulong eventIndex, IMessage @event);
+        Task PersistSnapshotAsync(string actorName, ulong eventIndex, IMessage snapshot);
         void Restart();
     }
 }
